+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 ## v0.2.0
 - Documentation
+- Define global output vite config